move draft models into core app
